# docs/swagger.yaml
definitions:
  models.Song:
    properties:
      group:
        type: string
      id:
        type: integer
      link:
        type: string
      releaseDate:
        type: string
      song:
        type: string
      text:
        type: string
    type: object
info:
  contact: {}
paths:
  /ping:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: pong
          schema:
            type: string
      summary: Health check
      tags:
      - Healthcheck
  /songs:
    get:
      parameters:
      - description: Song group
        in: query
        name: group
        type: string
      - description: Song title
        in: query
        name: title
        type: string
      - description: Song release date
        in: query
        name: release_date
        type: string
      - description: Song release date before
        in: query
        name: before
        type: string
      - description: Song release date after
        in: query
        name: after
        type: string
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Song'
            type: array
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Get songs
    post:
      parameters:
      - description: Song title
        in: body
        name: title
        required: true
        schema:
          type: string
      - description: Song title
        in: body
        name: group
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Created song
          schema:
            $ref: '#/definitions/models.Song'
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Create song
  /songs/{id}:
    delete:
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Delete song by id
    get:
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Song'
        "400":
          description: Bad request
          schema: {}
        "404":
          description: Not found
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Get song by id
    put:
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      - description: Song
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/models.Song'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Update song by id
swagger: "2.0"